Rename AppointmentsController to Account::AppointmentsController.
