+++ app/views/layouts/_user.html.erb
@@ -8,8 +8,8 @@
         </a>
         <ul class="menu">
           <li class="menu-item"><%= link_to "My Loans & Holds", member_loans_path %></li>
-          <li class="menu-item"><%= link_to "Schedule an Appointment", new_appointment_path %></li>
-          <li class="menu-item"><%= link_to "My Appointments", appointments_path %></li>
+          <li class="menu-item"><%= link_to "Schedule an Appointment", new_account_appointment_path %></li>
+          <li class="menu-item"><%= link_to "My Appointments", account_appointments_path %></li>
           <li class="menu-item"><%= link_to "View Loan History", member_loan_history_path %></li>
           <li class="menu-item"><%= link_to "Account Settings", member_profile_path %></li>
           <li class="menu-item"><%= link_to('Logout', destroy_user_session_path, method: :delete) %></li>
