+++ app/views/layouts/application.html.erb
@@ -57,8 +57,8 @@
               <span class="nav-category">Member Links</span>
               <ul class="nav">
                 <li class="nav-item"><%= link_to "My Loans & Holds", member_loans_path %></li>
-                <li class="nav-item"><%= link_to "Schedule an Appointment", new_appointment_path %></li>
-                <li class="nav-item"><%= link_to "My Appointments", appointments_path %></li>
+                <li class="nav-item"><%= link_to "Schedule an Appointment", new_account_appointment_path %></li>
+                <li class="nav-item"><%= link_to "My Appointments", account_appointments_path %></li>
                 <li class="nav-item"><%= link_to "View Loan History", member_loan_history_path %></li>
                 <li class="nav-item"><%= link_to "Account Settings", member_profile_path %></li>
               </ul>
