+++ config/routes.rb
@@ -21,6 +21,7 @@ Rails.application.routes.draw do
 
   namespace :account do
     resources :holds, only: [:create]
+    resources :appointments, only: [:index, :new, :create]
   end
 
   get "member/history", to: "members#history", as: "member_loan_history"
@@ -101,7 +102,6 @@ Rails.application.routes.draw do
     get "/", to: "dashboard#index", as: "dashboard"
   end
 
-  resources :appointments, only: [:index, :new, :create]
 
   get "/s/:id", to: "short_links#show", as: :short_link
